[softwaremanager] prevent crash and check if hdd is present

+++ lib/python/Plugins/SystemPlugins/SoftwareManager/plugin.py
@@ -1731,7 +1731,10 @@ class UpdatePlugin(Screen):
 			os.system("opkg list-upgradable > /etc/last-upgrades-git.log")
 			if not os.system("grep 'skins-xionhdf' /etc/last-upgrades-git.log"):
 				print "Xion skin update = Yes"
-				open('/media/hdd/images/xionrestore','w').close()
+				if not os.path.exists("/media/hdd"):
+					print "no hdd present to save xionrestore file"
+				else:
+					open('/media/hdd/images/xionrestore','w').close()
 			else:
 				print "Xion skin update = No"
 				if os.path.exists('/media/hdd/images/hdfrestore'):
